% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Dij.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{data from Wikipedia}
\format{
a data frame with 18 lines and 3 variables:
\describe{
\item{v1}{v1...}
\item{V2}{v2...}
\item{W}{W...}
}
}
\usage{
wiki_graph
}
\description{
a dataset containing two variables
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s}
}
\keyword{datasets}
