% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Dij.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{data from Wikipedia}
\format{

}
\source{
\url{http://www.diamondse.info/}
}
\usage{
wiki_graph
}
\description{
a dataset containing two variables
}
\keyword{datasets}
